backend/requirements.txt
itsdangerous<2.0.1
flask==2.0.1
Flask-WTF==1.1.1
flask_restful==0.3.10
sqlalchemy==1.4.31
flask-sqlalchemy==2.4.4
requests==2.31.0
pandas==2.0.3
pyjwt==2.8.0
Jinja2<3.1
werkzeug==2.0.1
gunicorn==20.1.0
flask-socketio==5.3.3
google-search-results==2.4.2
gevent==23.9.1
gevent-websocket==0.10.1
ipython==8.13.0
psycopg2-binary==2.9.9
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.101.0
pydantic==1.10.9
beautifulsoup4==4.12.2
pytz==2023.3.post1
python-dateutil==2.8.2
azure-cognitiveservices-speech==1.32.1
python-json-logger==2.0.7
cryptography==41.0.4
tzlocal==5.0.1
pydub==0.25.1
grequests==0.7.0
stripe==6.6.0
pgvector==0.2.4
app-store-server-library==1.0.0
packaging==23.2
openai==1.3.7
tiktoken==0.4.0
boto3==1.28.56
configparser==6.0.0